+++ src/grabs/mod.rs
@@ -1,6 +1,9 @@
 pub mod move_grab;
 pub use move_grab::MoveSurfaceGrab;
 
+pub mod ocean_pan_grab;
+pub use ocean_pan_grab::OceanPanGrab;
+
 pub mod resize_grab;
 pub use resize_grab::ResizeSurfaceGrab;